Add files via upload
This is  an intuitive approach solution of Hackerrank Equal problem. It is not a dynamic programming solution nor a memoized solution. It doesn't include recursion either 
I have made use of static array instead of a vector. But you can adjust the code however you want.

This solution relies on the difference between maximum and minimum element in the array.
according to the difference , add 1 , 2 or 5 to all except the max value in the array.
For example let's look at this test case  [2, 2, 3, 7]
first off the difference between max and min element is 7 - 2 = 5
Since 5 >= 5 , add 5 to all but 7 , making the array [7, 7, 8 , 7]
Now let's find the difference again. It is 8 - 7 = 1
Since 1>= 1 and 1 < 2, add 1 to all but 8, making the arr [ 8 , 8 , 8 , 8]
Again let's find the difference between max and min. It will be 0. Because 8 - 8 = 0;
So When difference between min and max is zero, While loop stops.
I have kept track of the number of operation needed to make  element in the array even  by using 'count' veriable;
That might seem unintuitive. But the code definitely finds the min number of operation that is needed for any test case.